requirements.txt
fastapi

uvicorn[standard]

sqlalchemy

psycopg2-binary

alembic

pydantic[email]

black

PyJWT

python-decouple
